Refactors settings interfaces into a separate file
Moves settings-related interfaces to a dedicated file.

This improves code organization and avoids duplication
across components that use these interfaces.
